feat(variant): Route unsigned C++ integer constructors to Variant::UINT (#1652)
* feat(variant): Route unsigned C++ integer constructors to UINT

`Variant(uint8_t)`, `Variant(uint16_t)`, `Variant(uint32_t)`,
`Variant(uint64_t)`, and `Variant(Math::uint_alt_t)` now select the
`Variant::UINT` carrier instead of storing a reinterpreted `int64_t`.

Boundaries that declare their Variant type elsewhere keep the signed
carrier, because nothing in the engine declares `UINT` yet:

- `VariantInternal::make()` retags unsigned results, so every bound
  method result, property value, and FoundryExtension conversion keeps
  the `Variant::INT` its own `GetTypeInfo<T>` declaration names.
- `MethodBind::set_default_arguments()` retags an unsigned default whose
  parameter declares `INT`.
- `BitField<T>` is an unsigned container for signed enum flags.
- `PropertyInfo::usage`, `MethodInfo::flags`, and `Connection::flags`
  are documented as `int` in script-facing reflection.
- The `ArrayMesh` `_surfaces` schema and the JSON native
  `PackedByteArray` element list are persisted signed-integer contracts.
- The debugger wire protocol declares its thread id field as a signed
  integer.

Reflection masks reaching script therefore still compose with `int`
operands, and a scene carrying an unsigned-typed native property saves,
reloads, and instantiates without resource corruption.

* fix(variant): Close unsigned carrier leaks found in review

- `MethodBind::set_default_arguments()` no longer dereferences the
  argument type table, which extension binds only build in debug
  builds, so registering an extension method with an unsigned default
  in a release template stays safe.
- `JSON::stringify()` prints the unsigned carrier as a JSON number
  instead of falling through to the quoted-value fallback.
- Debugger selection ids are carried nominally into
  `TypedArray<uint64_t>`, whose declared element type is
  `Variant::INT`, so remote node selection still records ids.

* feat(io): Persist the unsigned carrier in binary resources

Binary resource saving had no `Variant::UINT` case, so an unsigned value
reaching a saved property wrote no type tag or payload while `save()`
still reported success, producing a resource that failed to load. Add
`VARIANT_UINT64` so the value keeps both its exact bits and its carrier,
matching the exact text persistence and marshalling already in place.

The script-facing rendering server dictionaries carry their unsigned
mesh and texture fields nominally, so surface format masks still compose
with the integer constants scripts name them with.

* fix(debugger): Keep remote inspection and time fields on the signed carrier

The scene debugger wire protocol declares the remote object id and each
property usage mask as signed integer fields and `deserialize()` refuses
a message that arrives on another carrier, so both are carried
nominally. The datetime dictionaries do the same for their `uint8_t`
month and weekday fields, which are documented as `int`.

Binary resources advertise format version 8 now that they can contain
the unsigned Variant kind, so an older reader rejects them as a newer
format instead of reporting them as corrupt.

* fix(extension): Carry native pointer wrappers on the signed carrier

`FoundryExtensionPtr` and `FoundryExtensionConstPtr` declare
`Variant::INT` and their internal accessors read the signed slot, so
their Variant conversion must not select the unsigned carrier. A typed
script override of a pointer-backed virtual otherwise rejects the
argument and the callback never runs.

Drop the now-redundant indirection in the unsigned text-persistence
helpers, whose comments described the pre-flip constructor behavior.

* fix(text): Keep script-facing glyph dictionary integers signed

The glyph dictionaries returned by the shaped-text APIs assign the
`uint8_t` count/repeat and `uint16_t` flags fields straight into a
Variant. Grapheme flag constants are signed, so a script masking
`glyph["flags"]` needs the signed carrier on both operands.

* fix(scene): Keep persisted and automation schema integers signed

The animation compression schema, the lightmap capture data schema, and
the automation trace payload assign unsigned C++ fields straight into a
Variant. All three are read back as signed integers, so they stay on the
signed carrier.

* fix(tests): Guard SceneDebugger test behind DEBUG_ENABLED

SceneDebuggerObject only exists inside the DEBUG_ENABLED block in
scene/debugger/scene_debugger.h, so a template_release build (which
defines neither DEBUG_ENABLED nor TOOLS_ENABLED) failed to compile the
unguarded test header.

+++ core/os/time.cpp
@@ -211,7 +211,7 @@ Dictionary Time::get_datetime_dict_from_unix_time(int64_t p_unix_time_val) const
 	UNIX_TIME_TO_YMD
 	Dictionary datetime;
 	datetime[YEAR_KEY] = year;
-	datetime[MONTH_KEY] = (uint8_t)month;
+	datetime[MONTH_KEY] = (int64_t)month;
 	datetime[DAY_KEY] = day;
 	// Unix epoch was a Thursday (day 0 aka 1970-01-01).
 	datetime[WEEKDAY_KEY] = Math::posmod(day_number + WEEKDAY_THURSDAY, 7);
@@ -226,7 +226,7 @@ Dictionary Time::get_date_dict_from_unix_time(int64_t p_unix_time_val) const {
 	UNIX_TIME_TO_YMD
 	Dictionary datetime;
 	datetime[YEAR_KEY] = year;
-	datetime[MONTH_KEY] = (uint8_t)month;
+	datetime[MONTH_KEY] = (int64_t)month;
 	datetime[DAY_KEY] = day;
 	// Unix epoch was a Thursday (day 0 aka 1970-01-01).
 	datetime[WEEKDAY_KEY] = Math::posmod(day_number + WEEKDAY_THURSDAY, 7);
@@ -248,13 +248,13 @@ String Time::get_datetime_string_from_unix_time(int64_t p_unix_time_val, bool p_
 	UNIX_TIME_TO_HMS
 	UNIX_TIME_TO_YMD
 	const String format_string = p_use_space ? "%04d-%02d-%02d %02d:%02d:%02d" : "%04d-%02d-%02dT%02d:%02d:%02d";
-	return vformat(format_string, year, (uint8_t)month, day, hour, minute, second);
+	return vformat(format_string, year, (int64_t)month, day, hour, minute, second);
 }
 
 String Time::get_date_string_from_unix_time(int64_t p_unix_time_val) const {
 	UNIX_TIME_TO_YMD
 	// Android is picky about the types passed to make Variant, so we need a cast.
-	return vformat("%04d-%02d-%02d", year, (uint8_t)month, day);
+	return vformat("%04d-%02d-%02d", year, (int64_t)month, day);
 }
 
 String Time::get_time_string_from_unix_time(int64_t p_unix_time_val) const {
@@ -266,7 +266,7 @@ Dictionary Time::get_datetime_dict_from_datetime_string(const String &p_datetime
 	PARSE_ISO8601_STRING(Dictionary())
 	Dictionary dict;
 	dict[YEAR_KEY] = year;
-	dict[MONTH_KEY] = (uint8_t)month;
+	dict[MONTH_KEY] = (int64_t)month;
 	dict[DAY_KEY] = day;
 	if (p_weekday) {
 		YMD_TO_DAY_NUMBER
@@ -285,7 +285,7 @@ String Time::get_datetime_string_from_datetime_dict(const Dictionary &p_datetime
 	EXTRACT_FROM_DICTIONARY
 	VALIDATE_YMDHMS("")
 	const String format_string = p_use_space ? "%04d-%02d-%02d %02d:%02d:%02d" : "%04d-%02d-%02dT%02d:%02d:%02d";
-	return vformat(format_string, year, (uint8_t)month, day, hour, minute, second);
+	return vformat(format_string, year, (int64_t)month, day, hour, minute, second);
 }
 
 int64_t Time::get_unix_time_from_datetime_dict(const Dictionary &p_datetime) const {
@@ -322,9 +322,9 @@ Dictionary Time::get_datetime_dict_from_system(bool p_utc) const {
 	OS::DateTime dt = OS::get_singleton()->get_datetime(p_utc);
 	Dictionary datetime;
 	datetime[YEAR_KEY] = dt.year;
-	datetime[MONTH_KEY] = (uint8_t)dt.month;
+	datetime[MONTH_KEY] = (int64_t)dt.month;
 	datetime[DAY_KEY] = dt.day;
-	datetime[WEEKDAY_KEY] = (uint8_t)dt.weekday;
+	datetime[WEEKDAY_KEY] = (int64_t)dt.weekday;
 	datetime[HOUR_KEY] = dt.hour;
 	datetime[MINUTE_KEY] = dt.minute;
 	datetime[SECOND_KEY] = dt.second;
@@ -336,9 +336,9 @@ Dictionary Time::get_date_dict_from_system(bool p_utc) const {
 	OS::DateTime dt = OS::get_singleton()->get_datetime(p_utc);
 	Dictionary date_dictionary;
 	date_dictionary[YEAR_KEY] = dt.year;
-	date_dictionary[MONTH_KEY] = (uint8_t)dt.month;
+	date_dictionary[MONTH_KEY] = (int64_t)dt.month;
 	date_dictionary[DAY_KEY] = dt.day;
-	date_dictionary[WEEKDAY_KEY] = (uint8_t)dt.weekday;
+	date_dictionary[WEEKDAY_KEY] = (int64_t)dt.weekday;
 	return date_dictionary;
 }
 
@@ -354,13 +354,13 @@ Dictionary Time::get_time_dict_from_system(bool p_utc) const {
 String Time::get_datetime_string_from_system(bool p_utc, bool p_use_space) const {
 	OS::DateTime dt = OS::get_singleton()->get_datetime(p_utc);
 	const String format_string = p_use_space ? "%04d-%02d-%02d %02d:%02d:%02d" : "%04d-%02d-%02dT%02d:%02d:%02d";
-	return vformat(format_string, dt.year, (uint8_t)dt.month, dt.day, dt.hour, dt.minute, dt.second);
+	return vformat(format_string, dt.year, (int64_t)dt.month, dt.day, dt.hour, dt.minute, dt.second);
 }
 
 String Time::get_date_string_from_system(bool p_utc) const {
 	OS::DateTime dt = OS::get_singleton()->get_datetime(p_utc);
 	// Android is picky about the types passed to make Variant, so we need a cast.
-	return vformat("%04d-%02d-%02d", dt.year, (uint8_t)dt.month, dt.day);
+	return vformat("%04d-%02d-%02d", dt.year, (int64_t)dt.month, dt.day);
 }
 
 String Time::get_time_string_from_system(bool p_utc) const {
